Remove default sorting order stuff
This turned out to not do what I wanted. I wanted each column to "know"
its default sorting order, so that e.g. clicking on a numerical column
would sort by descending order; this is how I interpreted "default".
Instead, DataTables' "default sorting order" means the order in which
the table is sorted when you first load the page, which is not useful
for me since I'm already doing that with python.